Quick and dirty unsafe queue API
